[refactor] Category Enumeration 파일 분리
